Former-commit-id: a2540c58641b01aad781f812b040652a1cbced1b

+++ back/src/main/java/site/chachacha/fitme/domain/product/dto/ProductSearchRequest.java
@@ -11,7 +11,7 @@ public class ProductSearchRequest {
 
     private Long lastId;
 
-    private Integer lastPopularityScore;
+    private Double lastPopularityScore;
 
     private Integer lastPrice;
